Separated Class method Loading and KDLoadingBlurView in another files
