+++ KDLoadingView/KDLoadingBlurView.swift
@@ -1,0 +1,13 @@
+//
+//  KDLoadingBlurView.swift
+//  KDLoadingView
+//
+//  Created by Rodrigo Soldi Lopes on 02/03/17.
+//  Copyright © 2017 Kaique Damato. All rights reserved.
+//
+
+import UIKit
+
+class KDLoadingBlurView: UIVisualEffectView {
+    public var loadingView : KDLoadingView?
+}
